Split definition into multiple files
